[기능개선][1,2단계] CarRunner에서 출력 로직을 RaceOutput으로 분리
